chore: Refactor Api Stack
Move most functions in the ApiStack to a separate
Api-construct. This construct is responsible for
creating the API Gateway and attached functions.

Since we move them to a different construct, this
will result in recreation of resources. Since all
of these are stateless, no impact is expected.

The API key will be recreated, but since no
production consumers of this API exist yet, this
isn't an issue.
